naming overflow fix
Limit monster names, object type names, and specific object names to
length of PL_PSIZ (63) as seemed to be originally intended.  tty's
getlin() limited the user input to 79 or 80 (for a normal 80-character
wide screen) so was unlikely to overflow, but curses allowed all three
name types to be up 255 characters long.  The core was appending that
to formatted monster or object without restricting the overall length.

The 3 or 4 pieces of the previous 3.7 fix have been combined into a
single commit.  Their were a lot of code mismatches needing manual
repair due to 3.7's switch to C99, changing function prologs and
eliminating FDECL(), STATIC_OVL, and so forth.

This creates an empty fixes36.7 to keep everything in one place.  That
will need to be fixed up.